Abstract occurrence finding mechanism
The original implementation of LocalizationController would seek occurrences by
looking for strings using the 'find' function. This mechanism was incompatible
with certain types of occurrences. For example, HTMLTagOccurrences are not found
by searching for some string but by navigating through nested tags. To solve this
problem, the Occurrence class was equipped with a pure virtual function isFound()
which dictates how to search for a specific occurrence.