Replace Input file with Output
The script now creates an identical version of the input file that replaces the
original file. A temporary file called 'tmpFile' is created where all changes are
written, and, once the writing process is over, this temporary file replaces the
original file. If the program exits before the writing process is over, 'tmpFile'
does not replace the original file.