Add save line feature
The save line feature allows the saving of a line number for
situations when the occurrence handler can't handle a line
appropriately and it must be handled manually. The implementation
of the feature was done by modifying the Model interface and adding
setSavedLine and getSavedLine functions, hence why this is in
the cheapfix branch.